requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Werkzeug==2.3.7
requests==2.31.0
Flask-Mail==0.9.1
itsdangerous==2.1.2
python-dotenv==1.0.0